Move sundial integration into subfolder
