+++ .gitignore
@@ -2,4 +2,3 @@ catalog/node_modules
 catalog/training
 trainer/model.h5
 spec/data
-gs-build/
