+++ sundial/.gitignore
@@ -1,0 +1,2 @@
+gs-build/
+node_modules/
