+++ sundial/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "rollup": "^2.79.0"
+  }
+}
